Return 64-bit var for port_get_current_mtime

+++ FreeRTOSConfig.h
@@ -136,7 +136,7 @@
 /* Runtime stats definitions */
 /* TODO: use only for debugging */
 #define configUSE_STATS_FORMATTING_FUNCTIONS    1
-extern uint32_t port_get_current_mtime( void );
+extern uint64_t port_get_current_mtime( void );
 #define portCONFIGURE_TIMER_FOR_RUN_TIME_STATS()
 #define portGET_RUN_TIME_COUNTER_VALUE()    port_get_current_mtime()
